Finish icu independence
Some icu header files were still being included, because we were using icu's
definition of UChar to define OLECHAR. This was the only remaining use of icu.
We now define OLECHAR using uint16_t, which is what icu uses for UChar.
Because we use the same basic type as icu it's still possible for libcom
clients to use icu functions with OLECHAR strings.

Reviewed by: MarkS

svn-revision: 491